requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.3
django-crispy-forms==1.6.1
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
PyJWT==1.5.2
python-memcached==1.58
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tornado==4.5.1
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
